requirements.txt
streamlit==1.28.0
openai==1.3.0
pandas==2.1.0
python-dotenv==1.0.0
sqlite3
hashlib
json
datetime
typing